+++ lib/Predis/Configuration/ReplicationOption.php
@@ -1,0 +1,56 @@
+<?php
+
+/*
+ * This file is part of the Predis package.
+ *
+ * (c) Daniele Alessandri <suppakilla@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Predis\Configuration;
+
+use InvalidArgumentException;
+use Predis\Connection\MasterSlaveReplication;
+use Predis\Connection\ReplicationConnectionInterface;
+
+/**
+ * Configures an aggregate connection used for master/slave
+ * replication between multiple nodes.
+ *
+ * @author Daniele Alessandri <suppakilla@gmail.com>
+ */
+class ReplicationOption implements OptionInterface
+{
+    /**
+     * {@inheritdoc}
+     *
+     * @todo There's more code than needed due to a bug in filter_var()
+     *       as discussed here https://bugs.php.net/bug.php?id=49510.
+     */
+    public function filter(OptionsInterface $options, $value)
+    {
+        if ($value instanceof ReplicationConnectionInterface) {
+            return $value;
+        }
+
+        if (is_bool($value)) {
+            return $value ? $this->getDefault($options) : null;
+        }
+
+        if (!is_object($value) && null !== $asbool = filter_var($value, FILTER_VALIDATE_BOOLEAN, FILTER_NULL_ON_FAILURE)) {
+            return $asbool ? $this->getDefault($options) : null;
+        }
+
+        throw new InvalidArgumentException('Instance of Predis\Connection\ReplicationConnectionInterface expected');
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function getDefault(OptionsInterface $options)
+    {
+        return new MasterSlaveReplication();
+    }
+}